email: update .com -> .net
The .com domain has become a bit of a pain in the ass to deal with.
While I "own" it, managing it through Network Solutions has is difficult
and they continually give me the wrong transfer code to transfer out the
domain. Because it constantly feels at risk of being lost, I've
cancelled my plans of moving to it. All important emails will stay with
.net